requirements.txt
nose==1.3.7
pep8==1.7.1
requests==2.23.0
vcrpy==4.0.2